requirements.txt
openpyxl
pandas
numpy
pillow
supabase
python-dotenv